Fix multiple reg fault injection python wrapper type not match problem

+++ src/cpu/multi_register_fault_list.hh
@@ -14,9 +14,9 @@ class MultiRegisterFaultList : public SimObject
 private:
 	System* sys;
 	std::string faultLabel;
-	std::vector<uint64_t> bitPositionVector;
-	std::vector<uint64_t> faultRegisterVector;
-	std::vector<uint64_t> registerCategoryVector;
+	std::vector<uint32_t> bitPositionVector;
+	std::vector<uint32_t> faultRegisterVector;
+	std::vector<uint32_t> registerCategoryVector;
 	std::vector<uint32_t> faultStuckBitVector;
 	std::vector<uint64_t> tickVector;
 	std::vector<uint64_t> operationVector;
